Fix clippy

+++ src/columns/group.rs
@@ -35,7 +35,7 @@ impl Column for Group {
     fn add(&mut self, proc: &ProcessInfo) {
         let fmt_content = if let Some(ref status) = proc.curr_status {
             let gid = status.egid;
-            if let Some(group) = users::get_group_by_gid(gid as u32) {
+            if let Some(group) = users::get_group_by_gid(gid) {
                 format!("{}", group.name().to_string_lossy())
             } else {
                 format!("{}", gid)
