+++ src/columns/user_fs.rs
@@ -29,7 +29,7 @@ impl Column for UserFs {
     fn add(&mut self, proc: &ProcessInfo) {
         let fmt_content = if let Some(ref status) = proc.curr_status {
             let uid = status.fuid;
-            if let Some(user) = users::get_user_by_uid(uid as u32) {
+            if let Some(user) = users::get_user_by_uid(uid) {
                 format!("{}", user.name().to_string_lossy())
             } else {
                 format!("{}", uid)
